requirements.txt
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.6.15
charset-normalizer==2.0.12
Django==4.0.5
django-ckeditor==6.4.2
django-default-language==0.0.1
django-js-asset==2.0.0
django-modeltranslation==0.18.2
django-rosetta==0.9.8
idna==3.3
Pillow==9.1.1
polib==1.1.1
pytz==2022.1
requests==2.28.0
six==1.16.0
sqlparse==0.4.2
urllib3==1.26.9
